# internal/webui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "onderzeeer dashboard API",
    "version": "dev",
    "description": "HTTP API served by onderzeeerd when --web-listen is enabled. All paths are relative to this daemon.\n\nUse **Authorize** to enter the web token from the private token file shown in the daemon startup log. Enter only the token, without the Bearer prefix. Authorization is kept in memory and cleared when this page reloads.\n\nReads require authentication by default. With ONDERZEEER_WEB_PUBLIC_READ=true, GET operations accept anonymous requests; the served specification reflects that setting. An explicitly supplied invalid token is always rejected. Start and stop always require authentication and perform real control operations.\n\nThis reference and /openapi.json are readable without a token. They contain no queue data or credentials. The listener accepts localhost and loopback IP Host headers; wildcard listeners also accept remote literal IPs.\n\nThe CLI’s Unix-socket /v1/... control API is a separate transport and is not covered by this reference."
  },
  "servers": [
    {
      "url": "/",
      "description": "This onderzeeerd web listener"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/queues": {
      "get": {
        "operationId": "listQueues",
        "tags": [
          "Queues"
        ],
        "summary": "List registered queues and job counts",
        "description": "Includes queues belonging to stopped instances. A missing or unreadable database is reported in each queue’s database_state and error fields, with zero counts, without failing the whole response.",
        "responses": {
          "200": {
            "description": "List registered queues and job counts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Queues"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/instances": {
      "get": {
        "operationId": "listInstances",
        "tags": [
          "Instances"
        ],
        "summary": "List instances",
        "description": "Includes retained terminal instances by default. Set all=false to return active instances only (running or stopping).",
        "responses": {
          "200": {
            "description": "List instances.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Instances"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        },
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "required": false,
            "description": "Include terminal instances. Canonical values are true and false; Go boolean spellings such as 1 and 0 are also accepted.",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ]
      }
    },
...
    "/api/v1/queues/{queueID}/jobs/{jobID}": {
      "get": {
        "operationId": "getJob",
        "tags": [
          "Jobs"
        ],
        "summary": "Get a job with run and command history",
        "description": "Includes execution attempts and command metadata. Fetch stdout/stderr separately using each command ID. Persisted environment variables are not returned.",
        "responses": {
          "200": {
            "description": "Get a job with run and command history.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "description": "Queue or job not found (queue_not_found or job_not_found).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "job_not_found",
                    "message": "Job not found"
                  }
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/QueueUninitialized"
          },
          "503": {
            "$ref": "#/components/responses/QueueUnavailable"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/QueueID"
          },
          {
            "$ref": "#/components/parameters/JobID"
          }
        ]
      }
    },
...
    "/api/v1/instances/{instanceID}/stop": {
      "post": {
        "operationId": "stopInstance",
        "tags": [
          "Instances"
        ],
        "summary": "Stop an instance",
        "description": "Stops the instance and saves its desired state as stopped; its queue and execution history remain available.\n\nRequires a bearer token even when public reads are enabled, Content-Type: application/json, X-onderzeeer-Web: 1, and the JSON body {} (maximum 1024 bytes; unknown fields are rejected). Browser requests must originate from the same HTTP origin as the daemon.",
        "responses": {
          "200": {
            "description": "Stop an instance.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceResult"
                },
                "example": {
                  "instance": {
                    "id": "012345abcdef",
                    "name": "incoming",
                    "config_path": "/etc/onderzeeer/incoming.yaml",
                    "config_hash": "example-config-hash",
                    "database_path": "/var/lib/onderzeeer/incoming.db",
                    "state": "exited",
                    "desired_state": "stopped",
                    "created_at": "2026-09-26T12:00:00Z",
                    "started_at": "2026-09-26T12:00:00Z",
                    "finished_at": "2026-09-26T12:01:00Z"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "description": "Instance not found (not_found).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "Instance not found"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Control conflict (ambiguous_selector or not_active).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "not_active",
                    "message": "The instance is not active"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ShuttingDown"
          },
          "504": {
            "$ref": "#/components/responses/Timeout"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/InstanceID"
          },
          {
            "$ref": "#/components/parameters/MutationHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequest"
              },
              "example": {}
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The opaque web access token from the daemon token file (not a JWT)."
      }
    },
    "parameters": {
      "QueueID": {
        "name": "queueID",
        "in": "path",
        "required": true,
        "description": "Opaque ID from GET /api/v1/queues; not a queue name or filesystem path.",
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-f]{64}$"
        }
      },
      "JobID": {
        "name": "jobID",
        "in": "path",
        "required": true,
        "description": "Job ID from the queue job list.",
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid bearer token (unauthorized). An explicitly supplied invalid token is rejected even when public reads are enabled.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "unauthorized",
                "message": "A valid onderzeeer web token is required"
              }
            }
          }
        },
        "headers": {
          "WWW-Authenticate": {
            "description": "Bearer authentication challenge.",
            "schema": {
              "type": "string"
            },
            "example": "Bearer realm=\"onderzeeer-web\""
          }
        }
      },
      "InvalidRequest": {
        "description": "Invalid path/query parameter or mutation headers/body/origin (invalid_request).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "invalid_request",
                "message": "The request contains an invalid parameter, header, or body"
              }
            }
          }
        }
      },
...
      "QueueUninitialized": {
        "description": "The registered queue database has not been created yet (queue_uninitialized).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "queue_uninitialized",
                "message": "Queue has not been created yet"
              }
            }
          }
        }
      },
      "QueueUnavailable": {
        "description": "The queue database could not be read (queue_unavailable).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "queue_unavailable",
                "message": "The queue database could not be read"
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Queues": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "queues",
          "generated_at"
        ],
        "properties": {
          "queues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Queue"
            }
          },
          "generated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "queues": [],
          "generated_at": "2026-09-26T12:00:00Z"
        }
      },
      "Instances": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "instances",
          "generated_at"
        ],
        "properties": {
          "instances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Instance"
            }
          },
          "generated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "instances": [],
          "generated_at": "2026-09-26T12:00:00Z"
        }
      },
...
      "JobDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "job",
          "runs"
        ],
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Run"
            }
          }
        },
        "example": {
          "job": {
            "id": 1,
            "watch_name": "incoming",
            "path": "/data/incoming/report.txt",
            "status": "QUEUED",
            "attempts": 0,
            "max_retries": 3,
            "available_at": "2026-09-26T12:00:00Z",
            "created_at": "2026-09-26T12:00:00Z",
            "updated_at": "2026-09-26T12:00:00Z"
          },
          "runs": []
        }
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code."
              },
              "message": {
                "type": "string",
                "description": "Human-readable error detail."
              }
            }
          }
        }
      },
...
    }
  }
}